feat: add FileExtension, now it new result for main GetExtension, change name FileExtension2 -> ExtensionRoadMap & change structure to new types/names

+++ SpoofFileInfo/RoadMaps/ImageRoadMap.cs
@@ -6,11 +6,11 @@ public readonly record struct ImageRoadMap(
         int HeightOffset,
         int HeightSize,
         bool IsBigEndian,
-        int[][] SizeMaskOffset,
+        byte[][] SizeMaskOffset,
         int SizeOffset
     )
 {
-    private static readonly int[][] EmptySizeMask = [];
+    private static readonly byte[][] EmptySizeMask = [];
 
     public ImageRoadMap(
         int WitdhtOffset,
